+++ src/language/__tests__/schema-kitchen-sink.graphql
