+++ src/language/kinds.js
@@ -47,3 +47,15 @@ export const DIRECTIVE = 'Directive';
 export const NAMED_TYPE = 'NamedType';
 export const LIST_TYPE = 'ListType';
 export const NON_NULL_TYPE = 'NonNullType';
+
+// Type Definitions
+
+export const OBJECT_DEFINITION = 'ObjectDefinition';
+export const FIELD_DEFINITION = 'FieldDefinition';
+export const INPUT_VALUE_DEFINITION = 'InputValueDefinition';
+export const INTERFACE_DEFINITION = 'InterfaceDefinition';
+export const UNION_DEFINITION = 'UnionDefinition';
+export const SCALAR_DEFINITION = 'ScalarDefinition';
+export const ENUM_DEFINITION = 'EnumDefinition';
+export const ENUM_VALUE_DEFINITION = 'EnumValueDefinition';
+export const INPUT_OBJECT_DEFINITION = 'InputObjectDefinition';
