--- src/language/schema/README.md
@@ -1,10 +1,0 @@
-GraphQL Schema Language
------------------------
-
-The `graphql/language/schema` module is responsible for parsing and operating on
-the GraphQL schema definition language.
-
-```js
-import { ... } from 'graphql/language/schema'; // ES6
-var GraphQLSchemaLanguage = require('graphql/language/schema'); // CommonJS
-```
