--- src/language/schema/index.js
@@ -1,14 +1,0 @@
-/**
- *  Copyright (c) 2015, Facebook, Inc.
- *  All rights reserved.
- *
- *  This source code is licensed under the BSD-style license found in the
- *  LICENSE file in the root directory of this source tree. An additional grant
- *  of patent rights can be found in the PATENTS file in the same directory.
- */
-
-import * as Kind from './kinds';
-export { Kind };
-export { parseSchemaIntoAST } from './parser';
-export { printSchema } from './printer';
-export { visitSchema } from './visitor';
