--- src/language/schema/kinds.js
@@ -1,21 +1,0 @@
-/**
- *  Copyright (c) 2015, Facebook, Inc.
- *  All rights reserved.
- *
- *  This source code is licensed under the BSD-style license found in the
- *  LICENSE file in the root directory of this source tree. An additional grant
- *  of patent rights can be found in the PATENTS file in the same directory.
- */
-
-// Schema
-
-export const SCHEMA_DOCUMENT = 'SchemaDocument';
-export const TYPE_DEFINITION = 'TypeDefinition';
-export const FIELD_DEFINITION = 'FieldDefinition';
-export const INPUT_VALUE_DEFINITION = 'InputValueDefinition';
-export const INTERFACE_DEFINITION = 'InterfaceDefinition';
-export const UNION_DEFINITION = 'UnionDefinition';
-export const SCALAR_DEFINITION = 'ScalarDefinition';
-export const ENUM_DEFINITION = 'EnumDefinition';
-export const ENUM_VALUE_DEFINITION = 'EnumValueDefinition';
-export const INPUT_OBJECT_DEFINITION = 'InputObjectDefinition';
